+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Revision History
 
+## v ?.??.? - ?? ??? ????
+- SimpleShader no longer assumes VAOD and VBOD are the same value
+
 ## v 5.11.1 - 02 Sep 2025
 - Fixed bug that glGetProgramResource is only defined for version 4.3+