+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Revision History
 
-## v ?.??.? - ?? ??? ????
+## v 5.12.0 - 05 Sep 2025
 - SimpleShader no longer assumes VAOD and VBOD are the same value
 - SimpleShader supports deleting Vertex Arrays